Refactor object::data into common module
