+++ epp-client/src/domain/create.rs
@@ -2,10 +2,10 @@
 
 use epp_client_macros::*;
 
-use crate::epp::object::data::{
-    DomainAuthInfo, DomainContact, HostAttr, HostAttrList, HostList, HostObjList, Period,
+use crate::common::{
+    DomainAuthInfo, DomainContact, ElementName, EppObject, HostAttr, HostAttrList, HostList,
+    HostObjList, Period, StringValue,
 };
-use crate::epp::object::{ElementName, EppObject, StringValue};
 use crate::epp::request::Command;
 use crate::epp::response::CommandResponse;
 use crate::epp::xml::EPP_DOMAIN_XMLNS;
@@ -21,7 +21,7 @@ use serde::{Deserialize, Serialize};
 ///
 /// use epp_client::config::{EppClientConfig, EppClientConnection};
 /// use epp_client::EppClient;
-/// use epp_client::epp::object::data::DomainContact;
+/// use epp_client::common::DomainContact;
 /// use epp_client::domain::create::{EppDomainCreate, EppDomainCreateResponse};
 /// use epp_client::epp::generate_client_tr_id;
 ///
