+++ epp-client/src/domain/rgp/request.rs
@@ -2,8 +2,8 @@
 
 use epp_client_macros::*;
 
+use crate::common::{ElementName, EmptyTag, EppObject, Extension};
 use crate::domain::update::{DomainChangeInfo, DomainUpdateRequest, DomainUpdateRequestData};
-use crate::epp::object::{ElementName, EmptyTag, EppObject, Extension};
 use crate::epp::request::CommandWithExtension;
 use crate::epp::response::CommandResponseWithExtension;
 use crate::epp::xml::{EPP_DOMAIN_RGP_EXT_XMLNS, EPP_DOMAIN_XMLNS};
