+++ epp-client/src/epp.rs
@@ -1,6 +1,5 @@
 //! Types for EPP requests and responses
 
-pub mod object;
 pub mod request;
 pub mod response;
 pub mod xml;
