+++ epp-client/src/epp/xml.rs
@@ -7,7 +7,6 @@ use std::{error::Error, fmt::Debug};
 use crate::error;
 
 pub const EPP_XML_HEADER: &str = r#"<?xml version="1.0" encoding="UTF-8" standalone="no"?>"#;
-pub const EPP_XMLNS: &str = "urn:ietf:params:xml:ns:epp-1.0";
 
 pub const EPP_DOMAIN_XMLNS: &str = "urn:ietf:params:xml:ns:domain-1.0";
 pub const EPP_CONTACT_XMLNS: &str = "urn:ietf:params:xml:ns:contact-1.0";
