+++ epp-client/src/host/delete.rs
@@ -2,7 +2,7 @@
 
 use epp_client_macros::*;
 
-use crate::epp::object::{ElementName, EppObject, StringValue};
+use crate::common::{ElementName, EppObject, StringValue};
 use crate::epp::request::Command;
 use crate::epp::response::EppCommandResponse;
 use crate::epp::xml::EPP_HOST_XMLNS;
